webapp/backend/requirements.txt
absl-py==0.9.0
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.3.2
astunparse==1.6.3
attrs==19.3.0
awscli==1.16.274
backcall==0.2.0
bleach==3.1.5
boto3==1.10.14
botocore==1.13.14
Brotli==1.0.7
cachetools==4.1.0
certifi==2019.9.11
cffi==1.14.2
chalice==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==3.0
cycler==0.10.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
future==0.18.2
gast==0.3.3
gevent==20.6.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
httpie==1.0.3
idna==2.8
importlib-metadata==1.6.1
ipykernel==5.3.4
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.5.1
notebook==6.1.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2018.11.26
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.8
pycparser==2.20
pygame==1.9.6
Pygments==2.4.2
pylint==2.4.3
PyNaCl==1.4.0
pynamodb==4.2.0
pyOpenSSL==19.1.0
pyparsing==2.3.1
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2018.9
PyYAML==5.1.2
pyzmq==19.0.2
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
undecorated==0.3.0
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
